scripts/requirements.txt
requests
supabase
python-dotenv
pandas
xmltodict
